deleteTrip function Edit

+++ src/services/trip.service.js
@@ -40,12 +40,19 @@ const deleteTrip = async (id,res,tripId) => {
         const foundedTrip = await Trip.findById(tripId);
         if(foundedTrip){
             if(foundedTrip.Traveler.toString() == foundedTraveler._id.toString()){
+              if(foundedTrip.AcceptedRequests.length==0){
                const trip=await Trip.findByIdAndDelete(tripId);
                 foundedTraveler =  await Traveler.findByIdAndUpdate({_id:foundedTraveler._id},{$pull:{Trip:trip._id}},{new:true});
                 res.status(httpStatus.OK).json({
                     message: 'Trip deleted successfully',
                     trip
-                })}
+                }
+                )}else{
+                  res.status(httpStatus.INTERNAL_SERVER_ERROR).json({
+                    message: 'You can not delete this trip because there are accepted requests',
+                })
+                }
+              }
             else{
                 res.status(httpStatus.INTERNAL_SERVER_ERROR).json({
                     message: 'You are not allowed to delete this trip',
